+++ packages/event-listener/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @solid-primitives/event-listener
 
+## 2.2.12
+
+### Patch Changes
+
+- f03c47b0: Improve types of `createEventListenerMap`
+
 ## 2.2.11
 
 ### Patch Changes